tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tim Garvin", email = "tcgarvin@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.0",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
